Handler classes moved to handler.py module
